Moved support code for `rustc_mir::borrow_check` query out of `mod transform`.
